[consensus] Add randomness load test for forge
- Add DiceRoll entry point and dice_roll.move module for randomness workload
- Add realistic_env_max_load_randomness forge test with randomness-consuming
  transactions under realistic network conditions
- Include deadlock fix: eagerly send rand_tx in PipelinedBlock::set_randomness
  to prevent multi-block batch deadlock from PR #18699
- Rebuild prebuilt.mpb with dice_roll module

+++ crates/aptos-dkg/src/range_proofs/mod.rs
@@ -14,6 +14,6 @@ pub mod traits;
 #[derive(CanonicalSerialize)]
 pub struct PublicStatement<E: Pairing> {
     pub n: usize,
-    pub ell: usize,
+    pub ell: u8,
     pub comm: univariate_hiding_kzg::Commitment<E>,
 }
